added squarewave + frequency demo file
This file expands upon the previous example by implementing an input device(potentiometer) and using its value (which is then mapped to a more useful range of values) to change the delay in between pulses. This will change the frequency of the wave which the aux outputs. (essentially, pulse width modulation (PWM))